fruits-fax/backEnd/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
tensorflow==2.14.0
numpy==1.24.2
Pillow==9.4.0
Flask-CORS==4.0.0